Rename test definitions
