+++ src/cloudai/test_definitions/__init__.py
@@ -14,11 +14,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .chakra import ChakraReplayCmdArgs, ChakraReplayTestDefinition
-from .nccl import NCCLCmdArgs, NCCLTestDefinition
+from .chakra_replay import ChakraReplayCmdArgs, ChakraReplayTestDefinition
+from .nccl_test import NCCLCmdArgs, NCCLTestDefinition
 from .nemo_launcher import NeMoLauncherCmdArgs, NeMoLauncherTestDefinition
 from .sleep import SleepCmdArgs, SleepTestDefinition
-from .ucc import UCCCmdArgs, UCCTestDefinition
+from .ucc_test import UCCCmdArgs, UCCTestDefinition
 
 __all__ = [
     "ChakraReplayCmdArgs",
